Replacing processor from previous PR with new one + adapted configs accordingly
