+++ L1Trigger/L1THGCal/interface/backend/HGCalBackendStage1Processor.h
@@ -2,6 +2,7 @@
 #define __L1Trigger_L1THGCal_HGCalBackendStage1Processor_h__
 
 #include "L1Trigger/L1THGCal/interface/HGCalProcessorBase.h"
+#include "L1Trigger/L1THGCal/interface/HGCalAlgoWrapperBase.h"
 
 #include "DataFormats/L1THGCal/interface/HGCalTriggerCell.h"
 #include "DataFormats/L1THGCal/interface/HGCalCluster.h"
@@ -13,12 +14,12 @@ class HGCalBackendStage1Processor : public HGCalBackendStage1ProcessorBase {
 public:
   HGCalBackendStage1Processor(const edm::ParameterSet& conf);
 
-  void run(const std::pair<uint32_t, std::vector<edm::Ptr<l1t::HGCalTriggerCell>>>& fpga_id_tcs,
+  void run(const std::pair<uint32_t,std::vector<edm::Ptr<l1t::HGCalTriggerCell>>>& fpga_id_tcs,
            std::vector<edm::Ptr<l1t::HGCalTriggerCell>>& truncated_tcs,
-           const edm::EventSetup& es) override;
-
+           const edm::EventSetup& es);
 private:
-  std::unique_ptr<HGCalStage1TruncationImpl> truncation_;
+  std::unique_ptr<HGCalStage1TruncationWrapperBase> truncationWrapper_;
+  const edm::ParameterSet conf_;
 };
 
 #endif
